Python > 3.10 support; (only at installer/uninstaller)

+++ uninstall.sh
@@ -4,13 +4,16 @@
 # Uninstaller to Debian based Linux distros.
 ############################################################################################################
 
-# Installation paths
-MAIN_DIR=/home/$(whoami)/.UVMEnv
+### Installation paths ####
+REPO_PATH=$(pwd)
+HOME_DIR=/home/$(whoami)
+MAIN_DIR=$HOME_DIR/.UVMEnv
+VENV_DIR=$HOME_DIR/.UVMEnv_virtualenv
 REPOS_DIR=$MAIN_DIR/repos
 BASES_DIR=$MAIN_DIR/bases
 TOOLS_DIR=$MAIN_DIR/tools
 COMMAND=/usr/bin/uvmenv
-
+###########################
 
 ### Bash colors ####
 C_RED="\e[31m"
@@ -22,20 +25,53 @@ C_WHITE="\e[37m"
 C_N="\e[39m"
 ####################
 
+PY_VERSION=""
+
 REMOVE_ALL=0
+IS_PY10_OR_MINOR=0
+
 
 function main(){
+    # Firstly, get parameter 'all' value
     if [ "$1" == "all" ];then
         REMOVE_ALL=1
     fi
 
-    if [[ $REMOVE_ALL -eq 1 ]]; then
-        uninstallTools
-        uninstallPrerequisites
+    # Verify if UVMEnv is already uninstalled
+     if [ ! -d $MAIN_DIR ]; then
+        printWarning "UVMEnv is currently uninstalled"
+        exit 0
     fi
 
+    # UNINSTALLING PROCESS
+    # Remove simulators, python dependencies and system installation if 'all' is set (hard uninstalling)
+    if [[ $REMOVE_ALL -eq 1 ]]; then
+        uninstallSimulators
+
+        getPythonVersion
+        if [ $IS_PY10_OR_MINOR -eq 1 ]; then
+            uninstallPythonDependencies
+        else
+            rm -rf $VENV_DIR
+        fi
+        
+        # Autoremove system unused packages
+        sudo apt autoremove -y
+    fi
+    
+    # Remove UVMEnv install directory and command
     rm -rf $MAIN_DIR
+    if [ $IS_PY10_OR_MINOR -eq 0 ]; then
+        rm -rf $VENV_DIR
+    fi
     sudo rm $COMMAND
+
+    #Finally, show message
+    if [[ $REMOVE_ALL -eq 1 ]]; then
+        printInfo "UVMEnv has been HARD removed"
+    else
+        printInfo "UVMEnv has been removed"
+    fi
 }
 
 function printError(){
@@ -51,16 +87,9 @@ function printWarning(){
 }
 
 
-function uninstallPrerequisites(){
-    pip3 uninstall -y cocotb-coverage
-    pip3 uninstall -y cocotb 
-    pip3 uninstall -y pyuvm
-    pip3 uninstall -y colorama
-    pip3 uninstall -y pyfiglet
-    sudo apt autoremove -y
-}
 
-function uninstallTools(){
+
+function uninstallSimulators(){
     printInfo "############## Uninstalling jq... ##############"
     sudo apt purge --remove -y jq
 
@@ -82,6 +111,26 @@ function uninstallTools(){
     sudo rm -rf /usr/local/bin/verilator*
     sudo rm -rf /usr/local/bin/iverilog*
     sudo rm -rf /usr/local/bin/vvp
+}
+
+function getPythonVersion(){
+    local pyv_major=$(python3 --version | awk '{print $2}' | cut -d'.' -f1)
+    local pyv_minor=$(python3 --version | awk '{print $2}' | cut -d'.' -f2)
+    PY_VERSION=$pyv_major.$pyv_minor
+
+    if [ "$pyv_major" -gt 3 ] || { [ "$pyv_major" -eq 3 ] && [ "$pyv_minor" -gt 10 ]; }; then
+        IS_PY10_OR_MINOR=0
+    else
+        IS_PY10_OR_MINOR=1
+    fi
+}
+
+function uninstallPythonDependencies(){
+    python$PY_VERSION -m pip uninstall -y cocotb
+    python$PY_VERSION -m pip uninstall -y cocotb-coverage
+    python$PY_VERSION -m pip uninstall -y pyuvm
+    python$PY_VERSION -m pip uninstall -y pyfiglet
+    python$PY_VERSION -m pip uninstall -y colorama
 }
